+++ example/gcp_organisation/variable.tf
@@ -1,8 +1,15 @@
 variable "apikey"{
     type    = string
     default = ""
+    sensitive  = true
 }
 variable "region"{
     type    = string
     default = ""
+}
+
+variable "private_key"{
+    type       = string
+    default    = ""
+    sensitive  = true
 }