+++ example/report_config/main/provider.tf
@@ -1,7 +1,7 @@
 terraform {
   required_providers {
     conformity = {
-      version = "0.4.0"
+      version = "0.4.1"
       source  = "trendmicro/conformity"
     }
   }