+++ pkg/cloudconformity/get_report_config.go
@@ -9,7 +9,7 @@ func (c *Client) GetReportConfig(reportId string) (*ReportConfigDetails, error) 
 
 	reportConfigDetails := ReportConfigDetails{}
 
-	_, err := c.ClientRequest(Get{}, fmt.Sprintf("/v1/report-configs/%s", reportId), nil, "", &reportConfigDetails)
+	_, err := c.ClientRequest(Get{}, fmt.Sprintf("/report-configs/%s", reportId), nil, "", &reportConfigDetails)
 	if err != nil {
 		return nil, err
 	}