+++ pkg/cloudconformity/organization_external_id.go
@@ -3,7 +3,7 @@ package cloudconformity
 // allows you to get your organisation's external ID
 func (c *Client) GetExternalId() (string, error) {
 	externalId := externalIdData{}
-	_, err := c.ClientRequest(Get{}, "/v1/organisation/external-id/", nil, "", &externalId)
+	_, err := c.ClientRequest(Get{}, "/organisation/external-id/", nil, "", &externalId)
 	if err != nil {
 		return "", err
 	}