+++ pkg/cloudconformity/update_report_config.go
@@ -20,7 +20,7 @@ func (c *Client) UpdateReportConfig(reportId string, reportConfigPayload ReportC
 
 	log.Printf("[DEBUG] Conformity UpdateReportConfig request payload: %v\n", string(rb))
 
-	_, err = c.ClientRequest(Patch{}, fmt.Sprintf("/v1/report-configs/%s", reportId), strings.NewReader(string(rb)), "", &reportConfigDetails)
+	_, err = c.ClientRequest(Patch{}, fmt.Sprintf("/report-configs/%s", reportId), strings.NewReader(string(rb)), "", &reportConfigDetails)
 	if err != nil {
 		return "", err
 	}